Split project into two folders
Split the project into two folders, one with the old version and other with the new. The old version uses vue as an instance in a .js file, which is not the most common aproach. The new one will use a better aproach and will send data from one component to another